+++ Runtime/Groups/AudioGroup.cs
@@ -1,6 +1,5 @@
-using TMPro;
+using System;
 using UnityEngine;
-using UnityEngine.UI;
 using UnityEngine.Audio;
 
 namespace ActionCode.Audio
@@ -15,13 +14,8 @@ namespace ActionCode.Audio
         [SerializeField, Tooltip("The Volume Exposed Parameter name inside the AudioMixer asset.")]
         private string volumeParamName;
 
-        [Header("UI Components")]
-        [SerializeField, Tooltip("The local Toggle component.")]
-        private Toggle toggle;
-        [SerializeField, Tooltip("The local Slider controlling the Volume.")]
-        private Slider slider;
-        [SerializeField, Tooltip("The local Text component updated with Slider.")]
-        private TMP_Text sliderValue;
+        public event Action<float> OnVolumeChanged;
+        public event Action<bool> OnInteractableChanged;
 
         /// <summary>
         /// The Volume in linear range (0F -> 1F).
@@ -36,42 +30,26 @@ namespace ActionCode.Audio
             set
             {
                 mixer.SetFloat(volumeParamName, LinearToDecibel(value));
-                slider.SetValueWithoutNotify(value);
+                OnVolumeChanged?.Invoke(value);
+            }
+        }
 
-                var integerValue = (int)(value * 10);
-                sliderValue.text = integerValue.ToString("D2");
+        /// <summary>
+        /// Whether this Audio Group is interactable.
+        /// </summary>
+        public bool Interactable
+        {
+            get => enabled;
+            set
+            {
+                enabled = value;
+                Volume = enabled ? 1F : 0F;
+                OnInteractableChanged?.Invoke(enabled);
             }
         }
 
         private const float maxVolumeDB = 20F;
         private const float minVolumeDB = -80F;
-
-        private void Reset()
-        {
-            toggle = GetComponentInChildren<Toggle>();
-            slider = GetComponentInChildren<Slider>();
-            sliderValue = slider.transform.parent.GetComponentInChildren<TMP_Text>();
-        }
-
-        private void OnEnable()
-        {
-            toggle.onValueChanged.AddListener(HandleToggleValueChanged);
-            slider.onValueChanged.AddListener(HandleSliderValueChanged);
-        }
-
-        private void OnDisable()
-        {
-            toggle.onValueChanged.RemoveListener(HandleToggleValueChanged);
-            slider.onValueChanged.RemoveListener(HandleSliderValueChanged);
-        }
-
-        private void HandleToggleValueChanged(bool enabled)
-        {
-            slider.interactable = enabled;
-            Volume = enabled ? 1F : 0F;
-        }
-
-        private void HandleSliderValueChanged(float value) => Volume = value;
 
         private static float DecibelToLinear(float value) => Mathf.Pow(10F, value / maxVolumeDB);